[add] _env command
