+++ csml_engine/CSML/flow.csml
@@ -5,9 +5,13 @@ fn add(var1, var2):
     return 56
 
 start:
-    say "hello World"
-    // goto end
-    goto toto
+    say _env["amerelo"]
+    say _env["toto"]
+    say _env
+
+    say _env["hola"]
+    goto end
+    // goto toto
 
 toto:
     do vec = [1, 2]
