+++ csml_engine/examples/command_line.rs
@@ -60,6 +60,7 @@ fn init_bot() -> CsmlBot {
         custom_components: None,
         default_flow: "flowid".to_owned(),
         bot_ast: None,
+        env: None
     }
 }
