+++ csml_interpreter/examples/metadata.rs
@@ -30,6 +30,7 @@ fn main() {
         None,
         DEFAULT_FLOW_NAME,
         None,
+        None,
     );
 
     // Create an Event
