+++ csml_interpreter/tests/generic_component.rs
@@ -42,6 +42,7 @@ fn format_message(
             .to_owned())),
         DEFAULT_FLOW_NAME,
         None,
+        None,
     );
 
     interpret(bot, context, event, None)
