+++ csml_interpreter/tests/memory.rs
@@ -34,6 +34,7 @@ fn format_message(event: Event, context: Context, vector: &[&str]) -> MessageDat
         None,
         DEFAULT_FLOW_NAME,
         None,
+        None,
     );
 
     interpret(bot, context, event, None)
